+++ webpack/loader.cjs
@@ -1,6 +1,0 @@
-const contentSnapshot = require("./store");
-
-module.exports = function (content) {
-  contentSnapshot.set(this._module, content);
-  return content;
-};