+++ webpack/store.js
@@ -1,3 +1,0 @@
-const contentSnapshot = new WeakMap();
-
-module.exports = contentSnapshot;