+++ webpack/webpack.config.js
@@ -1,9 +1,0 @@
-const MapcssWebpackPlugin = require("./mod");
-
-/** @type import('webpack').Configuration */
-const config = {
-  mode: "development",
-  plugins: [new MapcssWebpackPlugin()],
-};
-
-module.exports = config;